Move ToSeq2 example under 1.23 build
