+++ example_test.go
@@ -817,23 +817,3 @@ func hash(input ...any) int {
 	fmt.Fprintln(hasher, input...)
 	return int(hasher.Sum32())
 }
-
-func Example_ToSeq2() {
-	nums := rill.FromSeq2(genPositive(40))
-	doubleNums := rill.Map(nums, 4, func(x int) (int, error) {
-		return x * x, nil
-	})
-	for val, err := range rill.ToSeq2(doubleNums) {
-		fmt.Println(val, err)
-	}
-}
-
-func genPositive(to int) iter.Seq2[int, error] {
-	return func(yield func(i int, err error) bool) {
-		for i := 1; i <= to; i++ {
-			if !yield(i, nil) {
-				return
-			}
-		}
-	}
-}
